Conflicts:
	README.mkd

+++ GpsMidGraph/de/ueller/gpsMid/mapData/ContainerTile.java
@@ -121,7 +121,7 @@ public class ContainerTile extends Tile {
 	    * coordinate and traverses that one first to check for
 	    * close by POI.
 	    */
-	public Vector getNearestPoi(boolean matchAnyPoi, byte searchType, float lat, float lon, float maxDist, CancelMonitorInterface cmi) {
+	public Vector getNearestPoi(boolean matchAnyPoi, short searchType, float lat, float lon, float maxDist, CancelMonitorInterface cmi) {
 		boolean t1closer;
 		Vector res;
 		Vector res2;
