+++ GpsMidGraph/de/ueller/gpsMid/mapData/FileTile.java
@@ -99,7 +99,7 @@ public class FileTile extends Tile implements QueueableTile {
 	}
 
 
-	public Vector getNearestPoi(boolean matchAnyPoi, byte searchType, float lat, float lon, float maxDist, CancelMonitorInterface cmi) {		
+	public Vector getNearestPoi(boolean matchAnyPoi, short searchType, float lat, float lon, float maxDist, CancelMonitorInterface cmi) {		
 		if (tile == null){
 			/**
 			 * This hole case is not correctly handled. The POI data
